requirements.txt
requests
python-dotenv
streamlit
google-generativeai
